Introduce special handling for 'deleted client' (#172)
* Take httr::refresh_oauth2.0() "as is"

* Start to morph towards something both more general and more Google-specific

Google-specific? Not worrying about `use_basic_auth`

More general? Stop filtering for specific oauth2.0 errors

* Don't try to delete a non-existent file

* Use rlang

* Make a function to detect legacy app name

* Note that I will ponder later

* Revise the signature

* Custom warning for 'deleted_client'

* Clarify comment and code

* Let's at least remove the refresh token

* Snapshot test the 'deleted_client' refresh failure

+++ R/oauth-cache.R
@@ -137,7 +137,7 @@ cache_load <- function(path) {
   }
 }
 
-cache_clean <- function(cache, pattern = "-calliope$") {
+cache_clean <- function(cache, pattern = gargle_legacy_app_pattern()) {
   # deletes an empty directory at the legacy cache location
   # new location implemented in v1.1.0
   # once we consider the transition done, this defer() can go away
@@ -165,6 +165,8 @@ cache_clean <- function(cache, pattern = "-calliope$") {
   file_delete(dat_tokens$filepath[dat_tokens$legacy])
   TRUE
 }
+
+gargle_legacy_app_pattern <- function() "-calliope$"
 
 # retrieve and insert tokens from cache -----------------------------------
 
@@ -200,8 +202,13 @@ token_remove_from_cache <- function(candidate) {
   cache_path <- candidate$cache_path
   if (is.null(cache_path)) return()
   token_path <- path(cache_path, candidate$hash())
-  gargle_debug(c("removing token from the cache:", "{.file {token_path}}"))
-  file_delete(token_path)
+  # when does token_path not exist?
+  # the first time a token fails to refresh it is removed on disk,
+  # but a package may still have it stored in its auth state
+  if (file_exists(token_path)) {
+    gargle_debug(c("Removing token from the cache:", "{.file {token_path}}"))
+    file_delete(token_path)
+  }
 }
 
 # helpers to compare tokens based on SHORTHASH_EMAIL ------------------------
